Fix usage of fetchTarball without sha256
In `nix/default.nix`, `fetchTarball` was being used to download
a fixed version of nixpkgs. However, no `sha256` value was given,
which means, according to the internet [1], that the fetched tarball
is cached for 1 hour only and then dowloaded again.

Instead of adding the `sha256` value, we use `fetchGit`, since then
we only need to know the commit hash, which makes it easier for people
to use when they want to bump nixpkgs. The downside is that the first
time the repo has to be fetched, it will be a bit slower.

[1] https://github.com/NixOS/nix/issues/2343