Add support for direnv
We add a `.envrc` file that puts all the environment variables
that `nix-shell` puts in scope into the current shell, every time
you `cd` to the directory.

While `direnv` has a standard support for nix, and `use nix` should
have sufficed, in practice, the time it takes for the environment
to be built is just long enough to be annoying. So instead, we use
the recipe from
[here](https://github.com/direnv/direnv/wiki/Nix#persistent-cached-shell).
In a nutshell, this caches the environment variables between calls,
and watches for changes in `default.nix` to invalidate the current
environment (without one having to do it manually).